apis/events,database: change "density" type to `Decimal(3,2)`
Update "density" field type to `Decimal(3,2)` to align with both the API
schema and the data warehouse schema. This ensures consistency in data
representation, eliminating the need for conversions when reading and
writing into the data warehouse.

+++ apis/datastore/events.go
@@ -69,7 +69,7 @@ var eventsMergeTable = warehouses.MergeTable{
 		{Name: "referrer_type", Type: types.Text()},
 		{Name: "screen_width", Type: types.Int(32)},
 		{Name: "screen_height", Type: types.Int(32)},
-		{Name: "screen_density", Type: types.Float(64)},
+		{Name: "screen_density", Type: types.Decimal(3, 2)},
 		{Name: "session_id", Type: types.Int(64)},
 		{Name: "session_start", Type: types.Boolean()},
 		{Name: "timezone", Type: types.Text()},
